Updated Workspace Settings component page to omit antd packages, refactored common components to generalize components usage.
- Changed Alert component to Callout
- Split the Loading components file to individual component files
- Created BaseFieldsForm and BaseAutoJoinForm, which take in props as primitive elements.
- Updated css retain consistent styling.
