+++ frontend/src/pages/WorkspaceSettings/WorkspaceSettings.module.css
@@ -10,3 +10,24 @@
 	margin-bottom: var(--size-xxLarge);
 	margin-top: var(--size-xxLarge);
 }
+
+.checkbox {
+	padding: 1px;
+	margin: 3px 0px;
+	border-radius: 4px;
+	svg {
+		width: 12px;
+		height: 12px;
+	}
+}
+@keyframes spin {
+	from {
+		transform: rotate(0deg);
+	}
+
+	to {
+		transform: rotate(360deg);
+	}
+}
+
+
